Add Avalonia.Browser project
