+++ src/StructuredLogViewer.Avalonia.Browser/Program.cs
@@ -1,0 +1,15 @@
+﻿using System.Runtime.Versioning;
+using Avalonia;
+using Avalonia.Web;
+using StructuredLogViewer.Avalonia;
+
+[assembly: SupportedOSPlatform("browser")]
+
+internal partial class Program
+{
+    private static void Main(string[] args) => BuildAvaloniaApp()
+        .SetupBrowserApp("out");
+
+    public static AppBuilder BuildAvaloniaApp()
+        => AppBuilder.Configure<App>();
+}
