+++ src/StructuredLogViewer.Avalonia/App.xaml.cs
@@ -15,6 +15,8 @@ namespace StructuredLogViewer.Avalonia
         {
             if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
                 desktop.MainWindow = new MainWindow();
+            else if (ApplicationLifetime is ISingleViewApplicationLifetime singleView)
+                singleView.MainView = new MainView();
 
             base.OnFrameworkInitializationCompleted();
         }
