+++ src/StructuredLogViewer.Avalonia/StructuredLogViewer.Avalonia.csproj
@@ -1,7 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>WinExe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
     <RuntimeIdentifiers>win7-x64;ubuntu.14.04-x64;osx.10.12-x64;osx-arm64</RuntimeIdentifiers>
     <ApplicationIcon>StructuredLogger.ico</ApplicationIcon>
@@ -36,11 +35,11 @@
 
   <ItemGroup>
     <PackageReference Include="Avalonia" Version="11.0-preview4" />
-    <PackageReference Include="Avalonia.Desktop" Version="11.0-preview4" />
     <PackageReference Include="Avalonia.Themes.Simple" Version="11.0-preview4" />
     <PackageReference Include="Avalonia.Diagnostics" Version="11.0-preview4" Condition=" '$(Configuration)' == 'Debug' " />
     <PackageReference Include="Avalonia.AvaloniaEdit" Version="11.0.0-preview2" />
     <PackageReference Include="Microsoft.DotNet.ILCompiler" Version="7.0.0-*" Condition="$(PublishNativeAot) == True" />
+    <PackageReference Include="XamlNameReferenceGenerator" Version="1.5.1" />
   </ItemGroup>
 
   <ItemGroup>
