+++ src/StructuredLogger/BinaryLogger/BinLogReader.cs
@@ -77,14 +77,14 @@ namespace Microsoft.Build.Logging.StructuredLogger
             // Use a producer-consumer queue so that IO can happen on one thread
             // while processing can happen on another thread decoupled. The speed
             // up is from 4.65 to 4.15 seconds.
-            var queue = new BlockingCollection<BuildEventArgs>(boundedCapacity: 5000);
-            var processingTask = System.Threading.Tasks.Task.Run(() =>
-            {
-                foreach (var args in queue.GetConsumingEnumerable())
-                {
-                    Dispatch(args);
-                }
-            });
+            // var queue = new BlockingCollection<BuildEventArgs>(boundedCapacity: 5000);
+            // var processingTask = System.Threading.Tasks.Task.Run(() =>
+            // {
+            //     foreach (var args in queue.GetConsumingEnumerable())
+            //     {
+            //         Dispatch(args);
+            //     }
+            // });
 
             int recordsRead = 0;
 
@@ -111,11 +111,12 @@ namespace Microsoft.Build.Logging.StructuredLogger
                 recordsRead++;
                 if (instance == null)
                 {
-                    queue.CompleteAdding();
+                    //queue.CompleteAdding();
                     break;
                 }
 
-                queue.Add(instance);
+                //queue.Add(instance);
+                Dispatch(instance);
 
                 if (progress != null && recordsRead % 1000 == 0 && stopwatch.ElapsedMilliseconds > 200)
                 {
@@ -126,7 +127,7 @@ namespace Microsoft.Build.Logging.StructuredLogger
                 }
             }
 
-            processingTask.Wait();
+            //processingTask.Wait();
 
             if (fileFormatVersion >= 10)
             {
@@ -313,10 +314,10 @@ namespace Microsoft.Build.Logging.StructuredLogger
         public override long Length => stream.Length;
 
         private long position;
-        public override long Position 
+        public override long Position
         {
-            get => position; 
-            set => throw new NotImplementedException(); 
+            get => position;
+            set => throw new NotImplementedException();
         }
 
         public override void Flush()
